requirements.txt
aiogram==2.6.1
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
fuzzywuzzy==0.18.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
numpy==1.18.1
pandas==1.0.1
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.12.0
python-Levenshtein==0.12.0
pytz==2019.3
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
yarl==1.4.2
